% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost.R
\name{AdaBoostClassifier}
\alias{AdaBoostClassifier}
\title{Adaboost classifier with quasi-randomized hidden layer}
\usage{
AdaBoostClassifier(
  obj,
  n_estimators = 10L,
  learning_rate = 0.1,
  n_hidden_features = 1L,
  reg_lambda = 0,
  reg_alpha = 0.5,
  activation_name = "relu",
  a = 0.01,
  nodes_sim = "sobol",
  bias = TRUE,
  dropout = 0,
  direct_link = FALSE,
  n_clusters = 2L,
  cluster_encode = TRUE,
  type_clust = "kmeans",
  col_sample = 1,
  row_sample = 1,
  seed = 123L,
  verbose = 1,
  method = "SAMME"
)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Adaboost classifier with quasi-randomized hidden layer
}
\examples{

library(datasets)

X <- as.matrix(iris[, 1:4])
y <- as.integer(iris[, 5]) - 1L

# obj <- sklearn$tree$DecisionTreeClassifier()
# obj2 <- AdaBoostClassifier(obj)
# obj2$fit(X, y)
# print(obj2$score(X, y))
# print(obj2$predict_proba(X))

}
