% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{CustomClassifier}
\alias{CustomClassifier}
\title{Custom classifier with quasi-randomized layer}
\usage{
CustomClassifier(
  obj,
  n_hidden_features = 5L,
  activation_name = "relu",
  a = 0.01,
  nodes_sim = "sobol",
  bias = TRUE,
  dropout = 0,
  direct_link = TRUE,
  n_clusters = 2L,
  cluster_encode = TRUE,
  type_clust = "kmeans",
  col_sample = 1,
  row_sample = 1,
  seed = 123L
)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
Custom classifier with quasi-randomized layer
}
\examples{

library(datasets)

X <- as.matrix(iris[, 1:4])
y <- as.integer(iris[, 5]) - 1L

obj <- Ridge2Classifier()
obj2 <- CustomClassifier(obj)
obj2$fit(X, y)
print(obj2$score(X, y))

}
