% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{GLMClassifier}
\alias{GLMClassifier}
\title{Generalized nonlinear models for Classification}
\usage{
GLMClassifier(
  n_hidden_features = 5L,
  lambda1 = 0.01,
  alpha1 = 0.5,
  lambda2 = 0.01,
  alpha2 = 0.5,
  family = "expit",
  activation_name = "relu",
  a = 0.01,
  nodes_sim = "sobol",
  bias = TRUE,
  dropout = 0,
  direct_link = TRUE,
  n_clusters = 2L,
  cluster_encode = TRUE,
  type_clust = "kmeans",
  type_scaling = c("std", "std", "std"),
  optimizer = ns$Optimizer(),
  seed = 123L
)
}
\description{
Parameters' description can be found at \url{https://techtonique.github.io/nnetsauce/}
}
