% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randombag.R
\name{RandomBagClassifier}
\alias{RandomBagClassifier}
\title{Bootstrap aggregating with quasi-randomized layer}
\usage{
RandomBagClassifier(
  obj,
  n_estimators = 10L,
  n_hidden_features = 1L,
  activation_name = "relu",
  a = 0.01,
  nodes_sim = "sobol",
  bias = TRUE,
  dropout = 0,
  direct_link = FALSE,
  n_clusters = 2L,
  cluster_encode = TRUE,
  type_clust = "kmeans",
  col_sample = 1,
  row_sample = 1,
  n_jobs = NULL,
  seed = 123L,
  verbose = 1L
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Bootstrap aggregating with quasi-randomized layer
}
\examples{

library(datasets)

X <- as.matrix(iris[, 1:4])
y <- as.integer(iris[, 5]) - 1L

obj <- sklearn$tree$DecisionTreeClassifier()
obj2 <- RandomBagClassifier(obj)
obj2$fit(X, y)
print(obj2$score(X, y))
print(obj2$predict_proba(X))

}
