% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge2.R
\name{Ridge2Classifier}
\alias{Ridge2Classifier}
\title{Multinomial logit, quasi-randomized classification model with 2 shrinkage parameters}
\usage{
Ridge2Classifier(
  n_hidden_features = 5L,
  activation_name = "relu",
  a = 0.01,
  nodes_sim = "sobol",
  bias = TRUE,
  dropout = 0,
  direct_link = TRUE,
  n_clusters = 2L,
  cluster_encode = TRUE,
  type_clust = "kmeans",
  col_sample = 1,
  row_sample = 1,
  lambda1 = 0.1,
  lambda2 = 0.1,
  seed = 123L,
  backend = c("cpu", "gpu", "tpu")
)
}
\arguments{
\item{backend}{}
}
\value{

}
\description{
Multinomial logit, quasi-randomized classification model with 2 shrinkage parameters
}
\examples{

# Example 1 -----

library(datasets)

X <- as.matrix(iris[, 1:4])
y <- as.integer(iris[, 5]) - 1L

obj <- Ridge2Classifier()
obj$fit(X, y)
print(obj$score(X, y))
print(obj$predict_proba(X))


# Example 2 -----

n <- 25
p <- 4
set.seed(123)
X <- matrix(rnorm(n * p), nrow = n, ncol = p)
y <- sample(c(0L, 1L), n, replace = TRUE)

obj2 <- Ridge2Classifier()
obj2$fit(X, y)
print(obj2$score(X, y))
print(obj2$predict_proba(X))

}
