% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvfl.R
\name{BayesianRVFLRegressor}
\alias{BayesianRVFLRegressor}
\title{Bayesian Random Vector Functional link network with 1 shrinkage parameter}
\usage{
BayesianRVFLRegressor(
  n_hidden_features = 5L,
  activation_name = "relu",
  a = 0.01,
  nodes_sim = "sobol",
  bias = TRUE,
  dropout = 0,
  direct_link = TRUE,
  n_clusters = 2L,
  cluster_encode = TRUE,
  type_clust = "kmeans",
  s = 0.1,
  sigma = 0.05,
  seed = 123L,
  backend = c("cpu", "gpu", "tpu")
)
}
\description{
Parameters' description can be found at \url{https://techtonique.github.io/nnetsauce/}
}
\examples{

set.seed(123)
n <- 50 ; p <- 3
X <- matrix(rnorm(n * p), n, p) # no intercept!
y <- rnorm(n)

(index_train <- base::sample.int(n = nrow(X),
                                 size = floor(0.8*nrow(X)),
                                 replace = FALSE))
X_train <- X[index_train, ]
y_train <- y[index_train]
X_test <- X[-index_train, ]
y_test <- y[-index_train]

obj <- BayesianRVFLRegressor(n_hidden_features = 5L)
print(obj$fit(X_train, y_train))
print(obj$score(X_test, y_test))

}
