% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts.R
\name{MTS}
\alias{MTS}
\title{Multivariate Time Series}
\usage{
MTS(
  obj,
  n_hidden_features = 5L,
  activation_name = "relu",
  a = 0.01,
  nodes_sim = "sobol",
  bias = TRUE,
  dropout = 0,
  direct_link = TRUE,
  n_clusters = 2L,
  cluster_encode = TRUE,
  type_clust = "kmeans",
  seed = 123L,
  lags = 1L,
  backend = c("cpu", "gpu", "tpu")
)
}
\description{
Parameters description can be found at \url{https://techtonique.github.io/nnetsauce/}
}
\examples{

# Example 1 -----

set.seed(123)
X <- matrix(rnorm(300), 100, 3)

obj <- sklearn$linear_model$ElasticNet()
obj2 <- MTS(obj)

obj2$fit(X)
obj2$predict()


# Example 2 -----

set.seed(123)
X <- matrix(rnorm(300), 100, 3)

obj <- sklearn$linear_model$BayesianRidge()
obj2 <- MTS(obj)

obj2$fit(X)
obj2$predict(return_std = TRUE)

}
